==> 1.0 Preparing Rating Data.ipynb <==
# Credit Ratings from Moodys

In this notebook I will get the raw data for Moody's credit rating.
These are XBRL files, therefore I have to take some steps before I can access the information.

# Importing libraries

# !pip install lxml

import pandas as pd
import numpy as np
from time import sleep
from random import randint
from bs4 import BeautifulSoup
import requests

# Libraries needed to read xblr documents
import csv
import pprint
import pathlib
import collections
# Parser for xblr, alternative to BeautifulSoup
import xml.etree.ElementTree as ET

# For indexing the different files
from os import listdir
from os.path import isfile, join

# Getting a list of the files

# Setting the path of the xblr files
rating_directory = pathlib.Path.cwd().joinpath('Raw Data').joinpath('XBRL-Files')
rating_directory

files_list = []

files_list = listdir(rating_directory)

len(files_list)
files_list[0]

# Creating a list of rating actions:

from IPython.display import clear_output

# This was for testing purposes
# Actions = []
# name = soup.select('OD > OBNAME')[0].text
# act_date = soup.select('ORD > RAD')[0].text
# rating = soup.select('ORD > R')[0].text
# action = [name,act_date,rating]
# Actions.append(action)

# Accessing all files
Actions = []
Counter = 0
Progress = 0.00
Progress_old = 0.00
Errorcount = 0
for filename in files_list:
    try:
        Counter += 1
        filepath = open(rating_directory.joinpath(filename))
        file = filepath.read()
        # Creating soup
        soup = BeautifulSoup(file, 'lxml')
        # Getting date from all ORD entries
        for entry in range(len(soup.select('ORD'))):
            # Appending the entries to our list
            name = soup.select('OD > OBNAME')[0].text
            # We convert commas to question marks to not get some weird errors when reading the csv, we will then change it back
            name = name.replace(',','?')
            act_date = soup.select('ORD > RAD')[entry].text
            rating = soup.select('ORD > R')[entry].text
            action = [name,act_date,rating]
            Actions.append(action)
            # Tracking progress
            Progress = round(Counter/len(files_list)*100,2)
            if Progress > Progress_old + 0.1:
                Progress_old = Progress
                clear_output()
                print(str(Progress)+'%')
                print('Number of errors so far: '+str(Errorcount))
                # Saving results every 0.1%
                actionframe = pd.DataFrame(Actions, columns = ['Company','Date','Rating'])
                actionframe.to_csv('Prepared Data/Rating_data'+str(Progress)+'.csv', index=False)
                # Reset the list
                Actions = []
    except:
        # If there is an error with processing a file, we will continues with the others
        Errorcount += 1
        pass
