==> 2.0 EDA and Data Cleaning.ipynb <==
# Importing needed libraries

# Basics
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
from datetime import date

# !pip install selenium

# For scraping
import pandas as pd
import numpy as np
from time import sleep
from random import randint
from bs4 import BeautifulSoup
import requests
from selenium import webdriver

# Loading the datasets

df_industries = pd.read_csv('./Raw Data/industries.csv', delimiter= ';')
df_companies = pd.read_csv('./Raw Data/us-companies.csv', delimiter= ';')
df_balance = pd.read_csv('./Raw Data/us-balance-quarterly.csv', delimiter= ';')
df_cashflow = pd.read_csv('./Raw Data/us-cashflow-quarterly.csv', delimiter= ';')
# df_shareprices = pd.read_csv('./Raw Data/us-shareprices-daily.csv', delimiter= ';')

# EDA & Cleaning

# Industries

display(df_industries.head())
display(df_industries.isna().sum())
display(df_industries.shape)

Defining the industry and sector according to an ID, all data seems to be complete and clean

# Companies

## First look

display(df_companies.head(3))
display(df_companies.isna().sum())
display(df_companies.shape)

## There is data missing for the industry ID's

# It seems that for some companies we do not have the indusry data for a lot of companies, we investiate.
df_companies[df_companies['IndustryId'].isna()].head(5)

Apparently the missing values are ranging from a very diverse set of industries.
I try to assign some of the missing values by deductingh from the company names.

## Tokenize
To better adress the different companies by name, I tokenize them

# We tokenize the company names, so we can search through all appearing words
from nltk.tokenize import word_tokenize

df_companies['tokens'] = df_companies['Company Name'].apply(word_tokenize)

df_companies.head(5)

## Functions for df_companies

# A function that returns one if one of our keywords is found
def finder(tokens):
    for i in tokens:
        if i.lower() in keywords:
            return 1
    return 0   

# A function, that displays the rows which contain keywords an have no industry ID 
# It also return a list of the corresponding indexes.

def nan_list(df,keywords):
    # Finding the companies with keywords
    df['current_search'] = 0
    df['current_search'] = df['tokens'].apply(finder)
    
    # Displaying them if they have no industrty-id
    display(df[ (df['current_search']==1) & (df['IndustryId'].isna()) ])
    
    # Saving the indexes
    indexlist = list(df[ (df['current_search']==1) & (df['IndustryId'].isna()) ].index)
    
    return indexlist

def set_industryID(industry, df, indexlist):
    
    # Getting the ID according to industry
    id = df_industries[df_industries['Industry']==industry]['IndustryId'].unique()[0]  
         
    # Setting the id for the specified indexes
    for row in indexlist:
        df.iloc[row,3] = id
    
    return df    

## ETF's & Funds

Since ETFs Funds and REITs do not usually issue debt notes or they have to be evaluated completely different, I will ecxlude them from this analysis.

keywords = ['etf','etc','fund','reit','trust']
df_companies['current_search'] = 0
df_companies['current_search'] = df_companies['tokens'].apply(finder)
# display(df_companies[df_companies['current_search']==1])
len(df_companies[df_companies['current_search']==1])

droppers1 = list(df_companies[df_companies['current_search']==1]['Ticker'])

df_companies.shape

# We drop the according rows
df_companies = df_companies[df_companies['current_search'] == 0]
df_companies.shape

## Scraping the missing data
Apparently nasdaq loads the sites content ad hoc using java scripts and normal get request won't work, therfore we use selenium to mimic a real browser access

# Importing the webdriver
from selenium import webdriver
from selenium.webdriver.common.by import By

# To wait until the driver finds the expected element, since it takes some time to load
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC

# Defning the path of the chromedriver
path = 'D:/Chromium/chromedriver'

# Setting up driver
# driver = webdriver.Chrome(path)

# Telling the driver to wait for up to 20 seconds if element is not found
# wait = WebDriverWait(driver, 20)

#### Create alist of company tickers with missing values

ticker_list = list(df_companies[df_companies['IndustryId'].isna()]['Ticker'])
len(ticker_list)

# Testing with wait function
# ticker = 'DOW'
# target_url = f"https://finance.yahoo.com/quote/%s/profile?p=%s" % (ticker,ticker)
# target_url

# driver.get(target_url)
# element = wait.until(EC.presence_of_element_located(('xpath', '//*[@id="Col1-0-Profile-Proxy"]/section/div[1]/div/div/p[2]/span[4]')))
# element.text

#### Scraping the missing data

Takes quite some time, so I saved the results as a pickle

# sectors = []
# for ticker in ticker_list:
#     # Setting url
#     target_url = f"https://finance.yahoo.com/quote/%s/profile?p=%s" % (ticker,ticker)
#     # Accessing the site
#     driver.get(target_url)    

#     # Getting the desired field:
#     try:
#         match = wait.until(EC.presence_of_element_located(('xpath', '//*[@id="Col1-0-Profile-Proxy"]/section/div[1]/div/div/p[2]/span[4]')))
#         sectors.append(match.text)
#         print(match.text)
#     except:
#         sectors.append('Not found')
#         print('Not found')
#     print(len(sectors))

# # Closing window
# driver.quit()

# # Saving Sectorlist
import pickle
# pickle.dump(sectors, open('sectors.p', 'wb'))

sectors = pickle.load(open('sectors.p','rb'))
# sectors

#### Assigning the scraped data to the company dataframe

# We translate the industry names to industry ID's
sectors_id = []
for i in range(len(sectors)):
    try:
        sectors_id.append(int(df_industries[df_industries['Industry'] == sectors[i]]['IndustryId']))
    except:
        # If the industry is not found, we append the Industry ID for 'Other'
        sectors_id.append('108004')

# We assign the industry names to a new column
df_companies.loc[df_companies['IndustryId'].isna(), ['IndustryId']] = sectors_id

# We got rid of all NaN's
df_companies['IndustryId'].isna().sum()

# We save the cleaned dataframe to our processed Data
df_companies.to_csv('Prepared Frames/company_data.csv', index=False)

#### Just keeping the important columns


df_companies = df_companies.drop(['SimFinId','tokens','current_search'], axis= 1)

# Balance sheets

df_balance = pd.read_csv('./Raw Data/us-balance-quarterly.csv', delimiter= ';')

display(df_balance.head(1))

## Dropping unneccessary columns

# We can already see, that there are several column that we won't need, so we drop them first:
# SimFinID : A site specific identifier
# Currency : We have everything in USD
# We will only need the Report Date if the dates
# The share count we decide to only used the Diluted since it better represents reality
# Treasury Stock is insignificant
# Total Liabilities and equity doesn't give us any information, that we don't have otheriwse
# Total Current Assets and Total assets are redundant

df_balance = df_balance.drop(['SimFinId','Currency','Publish Date','Restated Date','Shares (Basic)','Treasury Stock','Total Liabilities & Equity'], axis = 1)
df_balance.head(1)

## Dropping funds and ETF's
Since we earlier decided to exclude funds and ETF's we will now exclud ethem here as well.

for ticker in droppers1:
    df_balance = df_balance[df_cashflow['Ticker'] != ticker]

## Checking for NaN's

df_balance.isna().sum()

We will have to check column by column

df_balance[df_balance['Shares (Diluted)'].isna()]['Ticker'].unique()

# We see that the NaN's without share counts are split up into just 36 companies, we have a look, whart they are:
companylist = list(df_balance[df_balance['Shares (Diluted)'].isna()]['Ticker'].unique())

namelist = []
for company in companylist:
    try:
        namelist.append(str(df_companies[df_companies['Ticker']==company]['Company Name']))
    except:
        namlist.append(str(company)+' not found in companyframe')
namelist

# My guess, that we here have Funds that we won't include in our analysis did not hold.

# We check if all values are NAN for the share count
for company in companylist:
    print(len(df_balance[(df_balance['Ticker'] == company) & (df_balance['Shares (Diluted)'].isna())]) == len(df_balance[(df_balance['Ticker'] == company)]))

# Since we have no information about the sharecount for these companies and there are a lot of other values missing for them,
# I decide to drop them since they are just 36 out of over 3000
print(df_balance.shape)
for company in companylist:
    df_balance = df_balance[~(df_balance['Ticker']==company)]
df_balance.shape

## Systematic filling with mean value

Since we have many columns with a lot of missing valkues, wich make up more than 30% of our dataframe, and filling it with arbitrary values would cause a lot of error, we will drop them for now.

df_balance = df_balance.drop(['Inventories','Long Term Investments & Receivables','Short Term Debt','Long Term Debt'],axis = 1)
df_balance.shape

df_balance.isna().sum()

# The other columns, will fill the NaN's with the mean value for the corresponding company
def get_mean(company,column):
    return np.mean(df_balance[ (df_balance['Ticker']==company) & (df_balance[column].isna() == False) ][column])

# Columns to iterate thorugh
columns = ['Cash, Cash Equivalents & Short Term Investments','Accounts & Notes Receivable','Total Current Assets','Property, Plant & Equipment, Net',
          'Other Long Term Assets','Total Noncurrent Assets','Payables & Accruals','Total Noncurrent Liabilities','Share Capital & Additional Paid-In Capital',
          'Retained Earnings','Total Equity']

# Getting a list with all company tickers
companylist = list(df_balance['Ticker'].unique())

# setting mean for missing NaN
for column in columns:
    print(column)
    for company in companylist:
        df_balance.loc[ (df_balance[column].isna()) & (df_balance['Ticker']==company), [column]] = get_mean(company,column)

df_balance.isna().sum()

## Dropping companies with not enough data

# Apparently there are coulumns where no mean could be calculated. We see wich comnpanies are affected.
affected = []
for column in columns:
    affected.append(df_balance[df_balance[column].isna()]['Ticker'].unique())
# Making a list of the single entries instead of an array
droppers2 = []
for i in range(len(affected)):
    for x in affected[i]:
        droppers2.append(x)
# Removing duplicates
droppers2 = list(set(droppers2))
len(droppers2)

There are 336 companies affected for which we cannot get a value. We will have to drop them for now.

for ticker in droppers2:
    df_balance = df_balance[df_balance['Ticker'] != ticker]

df_balance.shape

df_balance.isna().sum()

# Saving:
df_balance.to_csv('Prepared Frames/company_data.csv', index=False)

# Saving the list of dropped companies for later:
tickers_to_drop = droppers1 + droppers2
len(tickers_to_drop)

# Cashflow Statements

df_cashflow = pd.read_csv('./Raw Data/us-cashflow-quarterly.csv', delimiter= ';')

display(df_cashflow.head(1))

df_cashflow.shape

## Dropping sorted out companies

# We already dropped some companies, that we cannot use from the balance sheet dataframe, we will drop them also here and then assess
# the situation
for ticker in tickers_to_drop:
    df_cashflow = df_cashflow[df_cashflow['Ticker'] != ticker]

df_cashflow.shape

## Dropping Unneccesary Columns

# We have again a lot of unneccessare columns, some of it even data, that we have in the balance sheet, so we drop them

to_drop = [
    'SimFinId','Currency','Report Date', 'Publish Date', 'Restated Date', 'Shares (Basic)',
    'Shares (Diluted)',
]

df_cashflow = df_cashflow.drop(to_drop, axis = 1)
print(df_cashflow.shape)
print(df_cashflow.columns)

## Checking NaNs

df_cashflow.isna().sum()

# We have some columns, that almost entirely consist out of NaNs, we drop them:
to_drop = [
    'Change in Accounts Receivable','Change in Inventories','Change in Accounts Payable', 'Change in Other'
]

df_cashflow = df_cashflow.drop(to_drop, axis = 1)
print(df_cashflow.shape)
print(df_cashflow.columns)

df_cashflow.isna().sum()

## Filling missing values with mean
We will try to implement the same strategy as before and fill the missing values with the means for the same company.

# The other columns, will fill the NaN's with the mean value for the corresponding company
def get_mean(company,column):
    return np.mean(df_cashflow[ (df_cashflow['Ticker']==company) & (df_cashflow[column].isna() == False) ][column])

# Columns to iterate thorugh
columns = ['Net Income/Starting Line','Depreciation & Amortization','Non-Cash Items','Change in Working Capital','Net Cash from Operating Activities',
          'Change in Fixed Assets & Intangibles','Net Change in Long Term Investment','Net Cash from Acquisitions & Divestitures',
          'Net Cash from Investing Activities','Dividends Paid','Cash from (Repayment of) Debt','Cash from (Repurchase of) Equity',
          'Net Cash from Financing Activities']

# Getting a list with all company tickers
companylist = list(df_cashflow['Ticker'].unique())

# setting mean for missing NaN
for column in columns:
    print(column)
    for company in companylist:
        df_cashflow.loc[ (df_cashflow[column].isna()) & (df_cashflow['Ticker']==company), [column]] = get_mean(company,column)

df_cashflow.isna().sum()

# We still have tweo columns wich make out nearle half our dataframe with NaNs, we drop them too.
df_cashflow = df_cashflow.drop(['Net Change in Long Term Investment','Dividends Paid'],axis = 1)
df_cashflow.isna().sum()

# For the column Net Cash from Acquisitions & Divestitures, 0 is a plausible value, since not every company makes an aquisition in every 
# quarter or sells part of its company. so we fill the missing values with 0.
# The same goes for Cash from (Repayment of) Debt ,Cash from (Repurchase of) Equity and Net Cash from financing activities.
df_cashflow.loc[df_cashflow['Net Cash from Acquisitions & Divestitures'].isna(),['Net Cash from Acquisitions & Divestitures']] = 0
df_cashflow.loc[df_cashflow['Cash from (Repayment of) Debt'].isna(),['Cash from (Repayment of) Debt']] = 0
df_cashflow.loc[df_cashflow['Cash from (Repurchase of) Equity'].isna(),['Cash from (Repurchase of) Equity']] = 0
df_cashflow.loc[df_cashflow['Net Cash from Financing Activities'].isna(),['Net Cash from Financing Activities']] = 0
df_cashflow.loc[df_cashflow['Net Cash from Investing Activities'].isna(),['Net Cash from Investing Activities']] = 0
df_cashflow.isna().sum()

# We look for which companies are making up the rest of the NaN
affected = []
columns = df_cashflow.columns
for column in columns:
    affected.append(df_cashflow[df_cashflow[column].isna()]['Ticker'].unique())
# Making a list of the single entries instead of an array
droppers3 = []
for i in range(len(affected)):
    for x in affected[i]:
        droppers3.append(x)
# Removing duplicates
droppers3 = list(set(droppers3))
len(droppers3)

# We can afford to loose these companies:
for ticker in droppers3:
    df_cashflow = df_cashflow[df_cashflow['Ticker'] != ticker]
df_cashflow.shape

# Since we have those companies still in the Balance dataframe we drop them there too:
for ticker in droppers3:
    df_balance = df_balance[df_balance['Ticker'] != ticker]
df_balance.shape

# Income Statements

df_income = pd.read_csv('./Raw Data/us-income-quarterly.csv', delimiter= ';')

df_income.columns

## Dropping unneccessary columns
The same as above

to_drop = [
    'SimFinId','Currency','Report Date', 'Publish Date', 'Restated Date', 'Shares (Basic)',
    'Shares (Diluted)',
]

df_income = df_income.drop(to_drop,axis = 1)

## Dropping companies as before
Since we already excluded some companies from our analysis, there is no point in keeping them here.

droppers_all = droppers1+droppers2+droppers3

for ticker in droppers_all:
    df_income = df_income[df_income['Ticker'] != ticker]

## Searching for NaNs

df_income.isna().sum()

# We still have some NaNs to clean up, two columns make up more than 50% NaNs, we drop them.
df_income = df_income.drop(['Research & Development','Depreciation & Amortization'], axis = 1)

# Also some columns are for extraordinary items, we assume that since they are missing there were no extraordinary events
# and fill them with 0.
extras = ['Abnormal Gains (Losses)','Net Extraordinary Gains (Losses)']
for column in extras:
    df_income[column] = 0

# For the rest we again fill with the companywise mean
# The other columns, will fill the NaN's with the mean value for the corresponding company
def get_mean(company,column):
    return np.mean(df_income[ (df_income['Ticker']==company) & (df_income[column].isna() == False) ][column])

df_income.isna().sum()

# Columns to iterate thorugh
columns = ['Revenue','Cost of Revenue','Gross Profit','Operating Expenses','Selling, General & Administrative',
           'Non-Operating Income (Loss)','Interest Expense, Net','Income Tax (Expense) Benefit, Net']

# Getting a list with all company tickers
companylist = list(df_income['Ticker'].unique())

# setting mean for missing NaN
for column in columns:
    print(column)
    for company in companylist:
        df_income.loc[ (df_income[column].isna()) & (df_income['Ticker']==company), [column]] = get_mean(company,column)

df_income.isna().sum()

At this point I will not delete any more companies and see if there is still a problem after choosing the companies for wich we have rating data.

#### Saving our dataframes again:


df_industries.to_csv('Prepared Frames/industry_data.csv', index=False)
df_companies.to_csv('Prepared Frames/company_data.csv', index=False)
df_balance.to_csv('Prepared Frames/balance_data.csv', index=False)
df_cashflow.to_csv('Prepared Frames/cashflow_data.csv', index=False)
df_income.to_csv('Prepared Frames/income_data.csv', index=False)
