==> 3.0 Merging Data.ipynb <==
# Here we merge our cleaned data

## Importing libraries

# Basics
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
from datetime import date

## Loading the Datasets

df_industries = pd.read_csv('Prepared Frames/industry_data.csv')
df_companies = pd.read_csv('Prepared Frames/company_data.csv')
df_balance = pd.read_csv('Prepared Frames/balance_data.csv')
df_cashflow = pd.read_csv('Prepared Frames/cashflow_data.csv')
df_income = pd.read_csv('Prepared Frames/income_data.csv')

display(df_industries.head(1))
display(df_companies.head(1))
display(df_balance.head(1))
display(df_cashflow.head(1))

## Merging

#### Cashflow and Balance Sheet
We will have to merge the cashflow and balance sheet data on the Ticker, Fiscal Year and Fiscal Period, to get the related reports
together.

print(df_balance.shape)
print(df_cashflow.shape)

df_statements = pd.merge(left = df_balance,
                                 right = df_cashflow,
                                 # We use outer to not loose Information, we will have to get rid of some NaNs later
                                 how = 'outer', 
                                 left_on = ['Ticker','Fiscal Year','Fiscal Period'], 
                                 right_on= ['Ticker','Fiscal Year','Fiscal Period'])
df_statements.shape

# df_statements.isna().sum()
# We created some rows with NaNs, we cannot get any additional information at this point and have to drop them.
df_statements = df_statements[df_statements['Report Date'].isna() == False]
df_statements.shape

#### Income data

print(df_statements.shape)
print(df_income.shape)

df_statements = pd.merge(left = df_balance,
                                 right = df_income,
                                 # We use outer to not loose Information, we will have to get rid of some NaNs later
                                 how = 'outer', 
                                 left_on = ['Ticker','Fiscal Year','Fiscal Period'], 
                                 right_on= ['Ticker','Fiscal Year','Fiscal Period'])
df_statements.shape

## Names and Industries
Now we will add the information from our companies dataframe, especially the name is needed, to assign the ratings later

df_main = pd.merge(left = df_companies,
                                 right = df_statements,
                                 # We use outer to not loose Information, we will have to get rid of some NaNs later
                                 how = 'right', 
                                 left_on = ['Ticker'], 
                                 right_on= ['Ticker'])
df_main.shape

# We will also add the Industry names, since it makes the analysis better understandable than the Industry IDs
df_main = pd.merge(left = df_main,
                                 right = df_industries,
                                 # We use outer to not loose Information, we will have to get rid of some NaNs later
                                 how = 'left', 
                                 left_on = ['IndustryId'], 
                                 right_on= ['IndustryId'])
df_main.shape

# df_main.isna().sum()
# No more NaNs were created

df_main.head(50)

## Saving the concatenated Dataframe

df_main.to_csv('Prepared Frames/main_data.csv', index=False)
