==> 4.0 Cleaning the Rating Data.ipynb <==
# Cleaning the rating Data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
from datetime import date

# For indexing the different files
from os import listdir
from os.path import isfile, join
import csv
import pprint
import pathlib
import collections

import time
from datetime import date

# Loading the credit rating actions:

#### Retrieving all the filepaths

# Path of the file directory
rating_directory = pathlib.Path.cwd().joinpath('Prepared Data')

# Gewtting list of all files
files_list = []
files_list = listdir(rating_directory)

rating_frames = []

# Getting all files as dataframes into a list
for file in files_list:
    filepath = open(rating_directory.joinpath(file))
    df = pd.read_csv(filepath)
    rating_frames.append(df)

# Merging all files
ratings_df = pd.concat(rating_frames, axis=0)

display(ratings_df.tail(20))
ratings_df.shape

# Changing back the question marks to commas
def convert(x):
    return x.replace('?',',')
ratings_df['Company'] = ratings_df['Company'].apply(convert)

# Converting the Date to datetime format
ratings_df['Date'] = pd.to_datetime(ratings_df['Date'])

## Data cleaning
Before we will take a look at the matches, we will clean up the other two columns of our dataframe.

#### Date column

# The good thing is, if there is anything but a date in the column, 
# the datetime function will throw an error.
ratings_df['Date'] = pd.to_datetime(ratings_df['Date'])

# Looks like the conversation here was successfull

#### Rating column
We have a look at the unique values in the rating column, those should ideally not be too many

len(ratings_df['Rating'].unique())
# I commented this out for not to bloat the notebook
# ratings_df['Rating'].unique()

The bad news: There are way too many unique values.
The good news: They all look like ratings, sometimes there are some values attached that do not belong there. We will try to get rid of them.

Most of the malformed rating follow the pattern, that they have a dot attached and then some characters.
We can help ourselfs here with a simple split function.

def remove_attach(rating):
    try:
        # We only take what is before the dot.
        res = rating.split('.')[0]
    except:
        # If there is no dot we leave the string as is.
        res = rating
    # Additionally we often see "(P)"  or "-PD" attached to our rating, we will get rid of that too.
    res = res.replace('(P)','')
    res = res.replace('-PD','')
    return res        

# We apply the cleaning function on our dataframe
ratings_df['Rating'] = ratings_df['Rating'].apply(remove_attach)
len(ratings_df['Rating'].unique())

# The rating that are relevant for us, since they are company credit ratings are now clean.
# We are talking about those, who Start with the letters A, B or C
ratings_df['Rating'].unique()

# Unnessesary ratings:
unnr = ['WR','NP', 'P-3', 'P-2', 'P-1','TR-1','MX-1','BR-1', 'BR-2', 'BR-4', 'SA-2', 'SA-1', 'BR-3', 'LB-1', 'LB-2',
       'LB-3', 'LB-4', 'MA-1', 'BO-1', 'MX-2', 'MX-3', 'MX-4', 'BO-3',
       'BO-2', 'TR-2', 'TR-3', 'TR-4', 'CZ-4', 'NG-1', 'NG-2', 'NG-4',
       'KE-1', 'MIG 1']

# Since we do not need the rating data, of municipalities states etc. we will drop them, to speed up
# the string searches we will do later.
ratings_df = ratings_df[ratings_df['Rating'].isin(unnr) == False ]

ratings_df['Rating'].unique()

## Saving the cleaned Ratings Data

ratings_df.to_csv('Prepared Frames/rating_data.csv', index=False)
