==> 5.1 Finding matches with Rating Data.ipynb <==
# Merging the Rating- to the Main Data
In this notebook we will merge the rating to the main Data.
The problem here are the differences in spelling between these two.

## Importing libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
from datetime import date

# !pip install Levenshtein
# !pip install fuzzywuzzy

# We will try to match the company names using different libraries and see wich method serves us best
import difflib
from difflib import SequenceMatcher
from fuzzywuzzy import fuzz

## Loading files

ratings_df = pd.read_csv('Prepared Frames/rating_data.csv')
df_companies = pd.read_csv('Prepared Frames/company_data.csv')

## Looking for matches
We will create a column for each method in our dataframe, where we will insert the closest match found with the ratings dataframe.

# Making a list of the companies in the ratings dataframe
r_companies = list(ratings_df['Company'].unique())
len(r_companies)

#### Making everything lowercase for better results


def makelow(x):
    return x.lower()

ratings_df['Company_lower'] = ratings_df['Company'].apply(makelow)

df_companies['Company_lower'] = df_companies['Company Name'].apply(makelow)

r_companies = list(ratings_df['Company_lower'].unique())

#### Difflib

# We want the matchs wo be very close, it is crucial to not get false matches,+
# while losing data is not ideal, but has to be expected in this case.
def getmatch(company):
    # We check if we have very close matches for the companies and return maximum of one value.
    try:
        return difflib.get_close_matches(company,r_companies,1,cutoff)[0]
    # If nothing is found, we return nothing, to get a NaN which will be easy for us to track
    except:
        pass

# Test
# The cutoff is just for testing purposes, we will use a way tighter one on our dataset.
cutoff = 0.5
getmatch('Barrick Gold')

# Applying it to our dataset

df_companies.columns

cutoff = 0.9
df_companies['difflib'] = df_companies['Company_lower'].apply(getmatch)

# Having a look how it worked
len(df_companies[df_companies['difflib'].isna()==False]['difflib'])
# 570 Matches, that looks better than the around 200 we had before.

# Check some examples on how good the matches are
# df_companies[df_companies['difflib'].isna()==False][['Company Name','difflib']].head(50)

The matches now look all correct, we had to increase the cutoff from initially 0.8 to 0.9 now.

#### Fuzzywuzzy
We try the same with the fuzzywuzzy library

from fuzzywuzzy import process

# We again define a function to get the closest matches, this time with the fuzzywuzzy library
def getfuzz(company):
    res,score = process.extractOne(company, r_companies, score_cutoff=1)
    if score < 1:
        pass
    else:
        return res   

print(getfuzz('Barrick Gold'))

The fuzzywuzzy results were way to vague and had a lot of mismatches, even with
the cutoff set to one. We cannot use this.

# This took a long time, so we save the results:
df_companies.to_csv('Prepared Frames/companies_approx_match.csv', index=False)

# Another attempt

When building the model I realized, that having 500 companies is not enough to train a model for all the rating classes.


We create yet another column, where we clean up the company names of a lot of unneccessary ballast and try to get some more matches like this.

def cleaner(s):
    # We will define alist of words that are often used within companynames but don't help to distinguish:
    bloat_list = ['incorporated','company','corporation','pharmaceuticals','limited','resources', 'incorporated','plc','inc','corp','co',
                  'group','holdings','.',',','-','ltd',"'",'&']
    
    s =  s.lower()
    # We remove the bloat from our targeted string
    for word in bloat_list:
        s = s.replace(word,'')
    # We return the cleaned phrase
    return s

# We create the new column
df_companies['reduced_name'] = df_companies['Company Name'].apply(cleaner)
df_companies.head(10)

# We do the same for the ratings
ratings_df['reduced_name'] = ratings_df['Company'].apply(cleaner)
ratings_df.head(10)

# We try to match again
r_companies = list(ratings_df['reduced_name'].unique())

df_companies['reduced_matches'] = df_companies['reduced_name'].apply(getmatch)

# Having a look how it worked
len(df_companies[df_companies['reduced_matches'].isna()==False]['reduced_matches'])

df_companies[df_companies['reduced_matches'].isna()==False][['Company Name','reduced_name','reduced_matches']].head(60)

df_companies[df_companies['reduced_matches'].isna()==False][['Company Name','reduced_name','reduced_matches']].tail(60)

# Adding the names of the matched companies, to our companies dataframe

# After looking through some examples, I am confident, that the matches are reliable.
# With each added word we were able to increase the number of matches for our dataframe.
# From initially 250 matched companies we went up to now 850
df_companies.to_csv('Prepared Frames/companies_approx_match.csv', index=False)

# We also have to save the ratings dataframe with the matched versions of the company names
ratings_df.to_csv('Prepared Frames/rating_data.csv', index=False)
