==> 5.2 Matching with Rating Data.ipynb <==
# We match our rating data to the main dataset

## Importing libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
from datetime import date

## Loading data

ratings_df = pd.read_csv('Prepared Frames/rating_data.csv')
df_companies = pd.read_csv('Prepared Frames/companies_approx_match.csv')
main_df = pd.read_csv('Prepared Frames/main_data.csv')

## Merging the found matches to the main Dataframe

main_df = pd.merge(left = main_df,
                                 right = df_companies[['Ticker','reduced_matches']],
                                 # We use outer to not loose Information, we will have to get rid of some NaNs later
                                 how = 'left', 
                                 left_on = ['Ticker'], 
                                 right_on= ['Ticker'])

## Assigning the ratings
We now have to assign the ratings. The difficulty is, that we don't have the exact dates to match, but have to match when the date in the main dataframe is greater than the one for the rating and have to overwrite it with a new rating in case there is one.

# We can only assign ratings where we have found the companies, 
# we create a new dataframe just with these entries
df_rated = main_df[main_df['reduced_matches'].isna()==False].copy()

# We sort the dataframe by dates
df_rated = df_rated.sort_values(by = ['Report Date'], ascending = True)
df_rated.head(3)

# We create a list of all unique companies in our ratings dataframe
companies = list(df_rated['reduced_matches'].unique())

len(companies)

#### Checking if we can match right

df_rated[df_rated['reduced_matches']==companies[0]].head(2)

ratings_df[ratings_df['reduced_name']==companies[0]]

ratings_df.columns

#### We have to make sure to overwrite only the greater dates with updated rating information
We will test it manually for the first company and do the rest with a function

# First create a sub-dataframe only with the ratings we are interested in
current_ratings = ratings_df[ratings_df['reduced_name']==companies[0]].sort_values(by = ['Date'], ascending = True).reset_index(drop = True)
current_ratings

for action in range(len(current_ratings)):
    # Getting the needed information
    company = current_ratings.loc[action,'reduced_name']
    date = current_ratings.loc[action,'Date']
    rating = current_ratings.loc[action,'Rating']
    
    # We need to adress the part of the dataframe where the company is the same and the date higher
    df_rated.loc[(df_rated['reduced_matches']==company)&(df_rated['Report Date']>=date),'Rating'] = rating

# Checking the result
# df_rated[df_rated['reduced_matches']==companies[0]]

# That worked correctly, we can apply it to our whole dataframe

def assign_ratings():
    # List of companies 
    companies = list(df_rated['reduced_matches'].unique())
    for company in companies:
        # Creating the subframe
        current_ratings = ratings_df[ratings_df['reduced_name']==company].sort_values(by = ['Date'], ascending = True).reset_index(drop = True)
        for action in range(len(current_ratings)):
            # Getting the needed information
            company = current_ratings.loc[action,'reduced_name']
            date = current_ratings.loc[action,'Date']
            rating = current_ratings.loc[action,'Rating']
             # We need to adress the part of the dataframe where the company is the same and the date higher
            df_rated.loc[(df_rated['reduced_matches']==company)&(df_rated['Report Date']>=date),'Rating'] = rating

# We apply the ratings
assign_ratings()

# Checking results
df_rated.isna().sum()

# We got almost everything rated
df_rated[df_rated['Rating'].isna()]

I looked through a few of them and realized, that the NaNs are created for the quarters where there is no rating available yet for the companies, so we drop those lines.

df_rated.shape

df_rated = df_rated.dropna(axis='rows')

df_rated.shape

# By doing this, all other NaNs vanished also
df_rated.isna().sum()

## Saving our finalized dataframe

df_rated.to_csv('Prepared Frames/rated.csv', index=False)
